refactor: Move the useSearchResult hook to Hooks folder
